+++ hypervisor/Makefile
@@ -176,7 +176,7 @@ C_SRCS += arch/x86/guest/vcpu.c
 C_SRCS += arch/x86/guest/vm.c
 C_SRCS += arch/x86/guest/vlapic.c
 C_SRCS += arch/x86/guest/vmtrr.c
-C_SRCS += arch/x86/guest/guest.c
+C_SRCS += arch/x86/guest/guest_memory.c
 C_SRCS += arch/x86/guest/vmcall.c
 C_SRCS += arch/x86/guest/vmsr.c
 C_SRCS += arch/x86/guest/instr_emul.c
