+++ hypervisor/arch/x86/guest/guest_memory.c
@@ -27,22 +27,6 @@ struct page_walk_info {
 	bool is_smep_on;
 };
 
-uint64_t vcpumask2pcpumask(struct acrn_vm *vm, uint64_t vdmask)
-{
-	uint16_t vcpu_id;
-	uint64_t dmask = 0UL;
-	struct acrn_vcpu *vcpu;
-
-	for (vcpu_id = 0U; vcpu_id < vm->hw.created_vcpus; vcpu_id++) {
-		if ((vdmask & (1UL << vcpu_id)) != 0UL) {
-			vcpu = vcpu_from_vid(vm, vcpu_id);
-			bitmap_set_nolock(vcpu->pcpu_id, &dmask);
-		}
-	}
-
-	return dmask;
-}
-
 enum vm_paging_mode get_vcpu_paging_mode(struct acrn_vcpu *vcpu)
 {
 	enum vm_cpu_mode cpu_mode;
@@ -439,58 +423,4 @@ int32_t copy_to_gva(struct acrn_vcpu *vcpu, void *h_ptr, uint64_t gva,
 	uint32_t size, uint32_t *err_code, uint64_t *fault_addr)
 {
 	return copy_gva(vcpu, h_ptr, gva, size, err_code, fault_addr, 0);
-}
-
-/**
- * @param[inout] vm pointer to a vm descriptor
- *
- * @retval 0 on success
- *
- * @pre vm != NULL
- * @pre is_sos_vm(vm) == true
- */
-void prepare_sos_vm_memmap(struct acrn_vm *vm)
-{
-	uint32_t i;
-	uint64_t attr_uc = (EPT_RWX | EPT_UNCACHED);
-	uint64_t hv_hpa;
-	uint64_t *pml4_page = (uint64_t *)vm->arch_vm.nworld_eptp;
-
-	const struct e820_entry *entry;
-	uint32_t entries_count = get_e820_entries_count();
-	const struct e820_entry *p_e820 = get_e820_entry();
-	const struct e820_mem_params *p_e820_mem_info = get_e820_mem_info();
-
-	dev_dbg(ACRN_DBG_GUEST, "sos_vm: bottom memory - 0x%llx, top memory - 0x%llx\n",
-		p_e820_mem_info->mem_bottom, p_e820_mem_info->mem_top);
-
-	if (p_e820_mem_info->mem_top > EPT_ADDRESS_SPACE(CONFIG_SOS_RAM_SIZE)) {
-		panic("Please configure SOS_VM_ADDRESS_SPACE correctly!\n");
-	}
-
-	/* create real ept map for all ranges with UC */
-	ept_mr_add(vm, pml4_page, p_e820_mem_info->mem_bottom, p_e820_mem_info->mem_bottom,
-			(p_e820_mem_info->mem_top - p_e820_mem_info->mem_bottom), attr_uc);
-
-	/* update ram entries to WB attr */
-	for (i = 0U; i < entries_count; i++) {
-		entry = p_e820 + i;
-		if (entry->type == E820_TYPE_RAM) {
-			ept_mr_modify(vm, pml4_page, entry->baseaddr, entry->length, EPT_WB, EPT_MT_MASK);
-		}
-	}
-
-	dev_dbg(ACRN_DBG_GUEST, "SOS_VM e820 layout:\n");
-	for (i = 0U; i < entries_count; i++) {
-		entry = p_e820 + i;
-		dev_dbg(ACRN_DBG_GUEST, "e820 table: %d type: 0x%x", i, entry->type);
-		dev_dbg(ACRN_DBG_GUEST, "BaseAddress: 0x%016llx length: 0x%016llx\n",
-			entry->baseaddr, entry->length);
-	}
-
-	/* unmap hypervisor itself for safety
-	 * will cause EPT violation if sos accesses hv memory
-	 */
-	hv_hpa = get_hv_image_base();
-	ept_mr_del(vm, pml4_page, hv_hpa, CONFIG_HV_RAM_SIZE);
 }
